# src/main/java/com/bootcamp/transactions/dto/Message.java
package com.bootcamp.transactions.dto;

import lombok.AllArgsConstructor;
import lombok.Data;

/**
 * Clase de Mensajes
 */
@Data
@AllArgsConstructor
public class Message {
	private String message;
}